requirements.txt
PyGithub==2.1.1
PyYAML==6.0.1
flake8==6.1.0
black==23.12.1
bandit==1.7.7
pytest==7.4.4
pytest-cov==4.1.0
mypy==1.8.0 